+++ src/main/java/org/tibid/dto/BidTicketDto.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 public class BidTicketDto extends BaseDto {
 	private long id;
 
-	private int userId;
+	private long userId;
 
-	private int bidOrderId;
+	private long bidOrderId;
 
 	private float price;
